Implement MVC architecture
