+++ src/can_explorer/__init__.py
@@ -1,0 +1,5 @@
+from can_explorer.controller import AppController
+from can_explorer.model import PlotModel
+from can_explorer.view import AppView
+
+_all__ = ["PlotModel", "AppView", "AppController"]
