+++ src/can_explorer/__main__.py
@@ -1,17 +1,18 @@
 import argparse
 import sys
 
-from can_explorer import app
+from can_explorer import AppController, AppView, PlotModel
 from can_explorer.resources.demo import demo_config
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--demo", action="store_true")
 args = parser.parse_args()
 
+controller = AppController(PlotModel(), AppView())
 
 if args.demo:
-    app.main(demo_config)
+    controller.run(demo_config)
 else:
-    app.main()
+    controller.run()
 
 sys.exit()
