+++ src/can_explorer/resources/__init__.py
@@ -1,7 +1,10 @@
 import pathlib
 import platform
 import uuid
+from random import randint
 from typing import Any, Final
+
+import can
 
 DIR_PATH: Final = pathlib.Path(__file__).parent
 
@@ -26,6 +29,16 @@ def generate_tag() -> int:
     Generate a random, unique tag.
     """
     return hash(uuid.uuid4())
+
+
+def generate_random_can_message() -> can.Message:
+    """
+    Generate a random CAN message.
+    """
+    message_id = randint(1, 25)
+    data_length = randint(1, 8)
+    data = (randint(0, 255) for _ in range(data_length))
+    return can.Message(arbitration_id=message_id, data=data)
 
 
 class Percentage:
