+++ src/can_explorer/resources/demo.py
@@ -4,17 +4,15 @@ from pathlib import Path
 
 import can.player
 
-from can_explorer import app, layout
-
 DEMO_FILE = Path(__file__).parent / "ic_sim.log"
 
 
-def demo_config() -> None:
+def demo_config(controller) -> None:
     # Use the virtual interface as default
-    layout.set_settings_interface_options(iterable=[""], default="virtual")
+    controller.view.set_settings_interface_options(iterable=[""], default="virtual")
 
     # Simulate the apply button press
-    app.settings_apply_button_callback(None, None, None)
+    controller.settings_apply_button_callback(None, None, None)
 
     # Play simulated logfile
     sys.argv = [sys.argv[0], "-i", "virtual", str(DEMO_FILE)]
